+++ packages/zowe-explorer/__tests__/__unit__/ZoweExplorerExtender.unit.test.ts
@@ -78,7 +78,7 @@ describe("ZoweExplorerExtender unit tests", () => {
             value: jest.fn(),
             configurable: true,
         });
-        Object.defineProperty(Profiles.getInstance(), "setConfigArray", {
+        Object.defineProperty(Profiles.getInstance(), "addToConfigArray", {
             value: jest.fn(),
             configurable: true,
         });