+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -7,10 +7,9 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 ### New features and enhancements
 
 - Added optional `getTag` function to `ZoweExplorerAPI.IUss` for getting the tag of a file on USS.
+- Added new API {ZE Extender MetaData} to allow extenders to have the metadata of registered extenders to aid in team configuration file creation from a view that isn't Zowe Explorer's. [#2394](https://github.com/zowe/vscode-extension-for-zowe/issues/2394)
 
 ### Bug fixes
-
-- Fixed the issue of getting the meta data of profile types for creating the config files by adding the setters and getters in the zowe-explorer-api. [#2394](https://github.com/zowe/vscode-extension-for-zowe/issues/2394)
 
 ## `2.11.0`