+++ packages/zowe-explorer/CHANGELOG.md
@@ -8,11 +8,11 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 
 - Added "Sort Jobs" feature for job nodes in Jobs tree view. [#2257](https://github.com/zowe/vscode-extension-for-zowe/issues/2251)
 - Added new edit feature on `Edit Attributes` view for changing file tags on USS [#2113](https://github.com/zowe/vscode-extension-for-zowe/issues/2113)
+- Added new API {ZE Extender MetaData} to allow extenders to have the metadata of registered extenders to aid in team configuration file creation from a view that isn't Zowe Explorer's. [#2394](https://github.com/zowe/vscode-extension-for-zowe/issues/2394)
 
 ### Bug fixes
 
 - Fixed submitting local JCL using command pallet option `Zowe Explorer: Submit JCL` by adding a check for chosen profile returned to continue the action. [#1625](https://github.com/zowe/vscode-extension-for-zowe/issues/1625)
-- Fixed the issue of getting the meta data of profile types for creating the config files by adding the setters and getters in the zowe-explorer-api. [#2394](https://github.com/zowe/vscode-extension-for-zowe/issues/2394)
 
 ## `2.11.0`